+++ pvi/Connection.py
@@ -63,7 +63,7 @@ class Connection():
         timeout = int(kwargs.get('timeout', 5 ))
         self._eventLoopIsRunning = False
         self._startTime = datetime.datetime.now()
-        self._pviTrialTimeCheck = datetime.timedelta(seconds=timeout +1)
+        self._pviManagerConnectionCheck = datetime.timedelta(seconds=timeout * 2 )
         self._objectsArranged = False
         self._pviObjects = []
         self._rootObject = PviObject(None, T_POBJ_TYPE.POBJ_PVI, '@Pvi')
@@ -229,7 +229,7 @@ class Connection():
         while run:  
             pviConnection.doEvents()  
         ```
-        """
+        """     
         wParam = WPARAM()
         lParam = LPARAM()
         hMsg = HANDLE()
@@ -310,9 +310,11 @@ class Connection():
                 # po = self.findObjectByLinkID(responseInfo.LinkID)
                 # print(f'{po.name} : nMode = {responseInfo.nMode}, nType = {responseInfo.nType}, ErrCode = {responseInfo.ErrCode}')                
         
-        if self._pviTrialTimeCheck:
-            if datetime.datetime.now() - self._startTime > self._pviTrialTimeCheck:
+        if self._pviManagerConnectionCheck:
+            if datetime.datetime.now() - self._startTime > self._pviManagerConnectionCheck:
+                PviDeinitialize()
                 raise PviError(12040)
+                
 
 
    # ----------------------------------------------------------------------------------